fixed issues with files not existance

+++ filter/cli.js
@@ -1,11 +1,13 @@
 module.exports = function(waw){
 	waw.exe('ng g p '+waw.path+' --module=common/common.module', function(){
-		waw.fs.unlinkSync(waw.base+'.pipe.spec.ts');
+		if (waw.fs.existsSync(waw.base+'.pipe.spec.ts')) {
+			waw.fs.unlinkSync(waw.base+'.pipe.spec.ts');
+		}
 		let ts = waw.fs.readFileSync(waw.params.template+'/filter.ts', 'utf8');
 		ts = ts.split('CNAME').join(waw.Name);
 		ts = ts.split('NAME').join(waw.name);
 		waw.fs.writeFileSync(waw.base+'.pipe.ts', ts, 'utf8');
-		let index = process.cwd() + '/src/app/filters/index.ts';
+		let index = process.cwd() + '/src/app/pipes/index.ts';
 		if (waw.fs.existsSync(index)) {
 			let index_exports = waw.fs.readFileSync(index, 'utf8') || '';
 			let code = "export { "+waw.Name+"Pipe } from './"+waw.name+".pipe';";
@@ -15,7 +17,7 @@ module.exports = function(waw){
 			}
 		}
 		let config = waw.fs.readFileSync(process.cwd() + '/src/app/common/common.module.ts', 'utf8');
-		let search = '/* filters */';
+		let search = '/* pipes */';
 		if(config && config.indexOf(search)>-1){
 			config = config.replace(search, search+'\n\t\t'+waw.Name+'Pipe,');
 			waw.fs.writeFileSync(process.cwd() + '/src/app/common/common.module.ts', config, 'utf8');
